Allow to add unmanaged tasks without allocating any resources if not needed.

+++ firmware/include/base_light.h
@@ -5,7 +5,7 @@
 class BaseLightDimmer : public CriticalTask<32> {
  public:
     static void initialize(uint16_t update_rate);
-    static void process(uint16_t cycle_count);
+    static void process(uint8_t cycle_count);
 
     static void set_brightness(uint8_t target);
     static void set_fade_target(uint8_t target);
